+++ MetaMorpheus/Test/UtilitiesTest/BioPolymerNotchFragmentIonComparerTest.cs
@@ -1,12 +1,11 @@
-﻿using EngineLayer;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using Omics.Fragmentation;
 using Proteomics;
 using Proteomics.ProteolyticDigestion;
 using System.Collections.Generic;
 using System.Reflection;
-using EngineLayer;
 using Omics.Modifications;
+using EngineLayer.SpectrumMatch;
 
 namespace Test.UtilitiesTest
 {
